Update README.md
SLMS Overview

The objective of the UN-REDD SLMS - Satellite Landcover Monitoring System - is to create an infrastructure that helps in the transparent dissemination of forest change, by processing and publishing forest (remote sensing) data, validated and open for public feedback of people in the monitored areas.

SLMS portal project Overview

This application is web map browser which act as a client of a OGC compliant infractructure. It is mostly used with Geoserver on the backend.

## Build and Run

From the root of the repo:

  ```
    #user@host$ cd src
    #user@host$ mvn clean install
  ```

the deploy the war in a tomcat6/java7 container.

A ready test environment can be setup running this [cookbook](https://github.com/slms4redd/chef-slms-portal) on a fresh Ubuntu 12.04.5 Virtual Machine or using Vagrant.